remove deprecated input finders
* cleans up specs
* removed plenty of deprecation warnings
* new installs will get default from formtastic form builder
