+++ lib/generators/templates/formtastic.rb
@@ -91,9 +91,8 @@
 
 # By creating custom input class finder, you can change how input classes  are looked up.
 # For example you can make it to search for TextInputFilter instead of TextInput.
-# See # TODO: add link # for more information
-# NOTE: this behavior will be default from Formtastic 4.0
-Formtastic::FormBuilder.input_class_finder = Formtastic::InputClassFinder
+# See https://github.com/justinfrench/formtastic/wiki/Custom-Class-Finders
+# Formtastic::FormBuilder.input_class_finder = Formtastic::InputClassFinder
 
 # Define custom namespaces in which to look up your Input classes. Default is
 # to look up in the global scope and in Formtastic::Inputs.
@@ -101,9 +100,8 @@ Formtastic::FormBuilder.input_class_finder = Formtastic::InputClassFinder
 
 # By creating custom action class finder, you can change how action classes are looked up.
 # For example you can make it to search for MyButtonAction instead of ButtonAction.
-# See # TODO: add link # for more information
-# NOTE: this behavior will be default from Formtastic 4.0
-Formtastic::FormBuilder.action_class_finder = Formtastic::ActionClassFinder
+# See https://github.com/justinfrench/formtastic/wiki/Custom-Class-Finders
+# Formtastic::FormBuilder.action_class_finder = Formtastic::ActionClassFinder
 
 # Define custom namespaces in which to look up your Action classes. Default is
 # to look up in the global scope and in Formtastic::Actions.
